Flip default behavior from loading CSV file, to querying for username
Previously, when run without any --csv or --user argument, we would default
to using a credentials.csv file - if existing - or failing.

Instead of defaulting to CSV mode, it is more useful for new users to be
prompted for their Garmin username/password, instead of requiring them
to either pass --user, or prepare a credentials.csv file beforehand.

This patch removes the --user option, and instead prompts the user for
both username and password, UNLESS -c/--csv was given at the command line,
in which case the given CSV file is used instead.